Fix bug where more mass would be assigned to first binary bin.
Before we were pulling mass from the working Mj array, not the original
so the first q in the list would get the most, then less to the second,
and even less to the third and so on. Now we pull the mass from the
original Mj array so each q should get the proper portion of mass.