Fix the bug where for multiple populations of binaries, fb was too large.
The conversion between the regular observational `fb` and the internal
`fb_ratio` quantity we use to calculate the amount of mass to shift
needed to be done before fb was split up into bins. This does however
make it non-trivial to specify your own distribution of `q`, so for now
I've switched the `shift_q` method to private with the intent to make a
wrapper which handles this properly.