Handle higher values of `fb`.
Previously with high values of `fb` bins with few stars would sometimes
go negative and we would just throw an exception complaining about `fb`.
Now we set that bin to zero and correct the binary bin by the right
amount to conserve mass and number. This does mean that at higher values
of `fb` we start to see more significant error (e.g. 5% error at `fb=45%`).
I think this is probably fine though as the error is small for reasonable values of `fb`.